+++ src/Commands/ShortScheduleRunCommand.php
@@ -8,7 +8,7 @@ use Spatie\ShortSchedule\ShortSchedule;
 
 class ShortScheduleRunCommand extends Command
 {
-    protected $signature = 'short-schedule:run {--pidfile=}';
+    protected $signature = 'short-schedule:run {--lifetime= : The lifetime in seconds of worker} {--pidfile= : pid file}';
 
     protected $description = 'Run the short scheduled commands';
 
@@ -21,6 +21,6 @@ class ShortScheduleRunCommand extends Command
             file_put_contents($pid_file, posix_getpid());
         }
 
-        (new ShortSchedule($loop))->registerCommands()->run();
+        (new ShortSchedule($loop))->registerCommands()->run($this->option('lifetime'));
     }
 }