+++ src/PendingShortScheduleCommand.php
@@ -4,13 +4,10 @@ namespace Spatie\ShortSchedule;
 
 use Closure;
 use Illuminate\Support\Arr;
-use Illuminate\Support\ProcessUtils;
 use Spatie\ShortSchedule\RunConstraints\BetweenConstraint;
 use Spatie\ShortSchedule\RunConstraints\EnvironmentConstraint;
 use Spatie\ShortSchedule\RunConstraints\RunConstraint;
 use Spatie\ShortSchedule\RunConstraints\WhenConstraint;
-use Symfony\Component\Console\Output\OutputInterface;
-use Illuminate\Console\Application;
 
 class PendingShortScheduleCommand
 {
@@ -24,11 +21,7 @@ class PendingShortScheduleCommand
 
     protected bool $evenInMaintenanceMode = false;
 
-    protected bool $runInBackground = false;
-
     protected array $constraints = [];
-
-    protected int $verbosity = OutputInterface::VERBOSITY_QUIET;
 
     public function everySecond(float $frequencyInSeconds = 1): self
     {
@@ -42,9 +35,13 @@ class PendingShortScheduleCommand
         return $this;
     }
 
-    public function command(string $artisanCommand):self
+    public function command(string $artisanCommand): self
     {
-        $this->command = Application::formatCommandString($artisanCommand);
+        if (class_exists($artisanCommand)) {
+            $artisanCommand = app($artisanCommand)->getName();
+        }
+
+        $this->command = PHP_BINARY . " artisan {$artisanCommand}";
 
         return $this;
     }
@@ -115,27 +112,13 @@ class PendingShortScheduleCommand
         return $this;
     }
 
-    public function verbose(): self
+    public function getOnOneServer(): bool
     {
-        $this->verbosity = OutputInterface::VERBOSITY_NORMAL;
-
-        return $this;
-    }
-
-    public function runInBackground(): self
-    {
-        $this->runInBackground = true;
-
-        return $this;
+        return $this->onOneServer;
     }
 
     public function cacheName(): string
     {
         return 'framework'.DIRECTORY_SEPARATOR.'schedule-'.sha1($this->frequencyInSeconds.$this->command);
-    }
-
-    public function cacheNameOnOneServer(): string
-    {
-        return 'framework'.DIRECTORY_SEPARATOR.'schedule-'.sha1($this->frequencyInSeconds.$this->command.'onOneServer');
     }
 }