+++ src/ShortScheduleServiceProvider.php
@@ -3,7 +3,6 @@
 namespace Spatie\ShortSchedule;
 
 use Illuminate\Support\ServiceProvider;
-use Spatie\ShortSchedule\Commands\ShortScheduleFinishCommand;
 use Spatie\ShortSchedule\Commands\ShortScheduleRunCommand;
 
 class ShortScheduleServiceProvider extends ServiceProvider
@@ -13,7 +12,6 @@ class ShortScheduleServiceProvider extends ServiceProvider
         if ($this->app->runningInConsole()) {
             $this->commands([
                 ShortScheduleRunCommand::class,
-                ShortScheduleFinishCommand::class
             ]);
         }
     }